Changed SPARK dependency to 1.1 | Integrated Interpreter and IScalaInterpreter | Fixed test code
